pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-studio"
version = "0.1.0"
description = "Cross-platform desktop skeleton for an AI-assisted automation workspace."
readme = "README.md"
requires-python = ">=3.11,<3.14"
authors = [
  { name = "Codex" }
]
dependencies = [
  "PySide6>=6.9,<7",
  "fastapi>=0.115,<1",
  "uvicorn[standard]>=0.34,<1",
  "httpx>=0.28,<1",
  "Pillow>=11,<12",
  "mss>=9,<11",
  "pydantic-settings>=2.10,<3",
  "python-dotenv>=1.1,<2"
]

[project.optional-dependencies]
automation = [
  "pynput>=1.7,<2"
]
dev = [
  "pytest>=8.3,<9",
  "ruff>=0.11,<1"
]
perception = [
  "rapidocr_onnxruntime>=1.4,<2; python_version < '3.13'"
]

[project.scripts]
agent-studio = "agent_studio.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-p no:cacheprovider"
pythonpath = ["src"]
testpaths = ["tests"]
